@@ -1,20 +1,15 @@
-/* global require, module */
-
-const
-    React = require('react'),
-    Router = require('react-router'),
-    RouteHandler = Router.RouteHandler;
+import React from 'react';
 
 /**
- * React component AppLayout
- *
- * @class AppLayout
+ * @class
+ * @memberof module:Demo
+ * @classdesc AppLayout is a basic layout for the demo adding a div.app around its children.
  */
 const AppLayout = React.createClass({
     render: function() {
         return (
             <div className='app'>
-                <RouteHandler />
+                {this.props.children}
             </div>
         );
     }
