@@ -1,31 +1,35 @@
-/* global require, module */
-
-const
-    React = require('react'),
-    Router = require('react-router'),
-    Route = Router.Route,
-    CompositeRoutes = require('reactator/routes/CompositeRoutes'),
-    MainRoutes = require("./MainRoutes.js"),
-    ProfileRoutes = require('./ProfileRoutes.js'),
-    AppLayout = require("../components/AppLayout");
+import React from 'react';
+import {Route} from 'react-router';
+import CompositeRoutes from 'reactator/routes/CompositeRoutes';
+import MainRoutes from './MainRoutes.js';
+import ProfileRoutes from './ProfileRoutes.js';
+import AppLayout from '../components/AppLayout.js';
 
 /**
- * @class DemoRoutes
+ * @class
+ * @memberof module:Demo
+ * @classdesc DemoRoutes is an example of {@link module:Reactator.CompositeRoutes} for the demo app.
  */
 class DemoRoutes extends CompositeRoutes {
     constructor() {
         super([
-                new MainRoutes(),
-                new ProfileRoutes()
-            ]);
+            new MainRoutes(),
+            new ProfileRoutes()
+        ]);
     }
 
+    /**
+     * Provides the composite react-router routes for the demo.
+     *
+     * @return {Route} composite route
+     * @inheritdoc
+     */
     getRoutes() {
         return (
-                <Route handler={AppLayout}>
-                    {super.getRoutes()}
-                </Route>
-            );
+            <Route component={AppLayout}>
+                {super.getRoutes()}
+            </Route>
+        );
     }
 }
 
