@@ -1,26 +1,26 @@
-/* global require, module */
-
-const
-    React = require('react'),
-    Router = require('react-router'),
-    Route = Router.Route,
-    DefaultRoute = Router.DefaultRoute,
-    Routes = require('reactator/routes/Routes'),
-    BasicLayout = require("reactator/components/BasicLayout"),
-    ProfileComponent = require("../components/profile/ProfileComponent");
+import React from 'react';
+import {Route, IndexRoute} from 'react-router';
+import Routes from 'reactator/routes/Routes';
+import BasicLayout from 'reactator/components/BasicLayout';
+import ProfileComponent from '../components/profile/ProfileComponent.js';
 
 /**
- * React Router definition for profile app
- *
- * @class ProfileRoutes
+ * @class
+ * @memberof module:Demo
+ * @classdesc ProfileRoutes is an example of {@link module:Reactator.Routes} for the demo app's profile page.
  */
 class ProfileRoutes extends Routes {
+    /**
+     * Provides the routes for the profile component of the demo app.
+     * @return {Route} the route representing the profile component
+     * @inheritdoc
+     */
     getRoutes() {
         return (
-                <Route key="profile" name="profile" handler={BasicLayout}>
-                    <DefaultRoute handler={ProfileComponent} />
-                </Route>
-            );
+            <Route key="profile" name="profile" path="profile" component={BasicLayout}>
+                <IndexRoute component={ProfileComponent} />
+            </Route>
+        );
     }
 }
 
