@@ -1,48 +1,47 @@
-/* global require, document */
+/** @module Reactator */
 
-const
-    React = require('react'),
-    Router = require('react-router'),
-    AppDispatcher = require('./dispatcher/AppDispatcher.js'),
-    MetaDataComponent = require('./components/MetaDataComponent.js');
+import React from 'react';
+import ReactDOM from 'react-dom';
+import {Router, Route, browserHistory} from 'react-router';
+import AppConstants from './constants/AppConstants.js';
+import RootRouteComponent from './components/RootRouteComponent.js';
+import MetaDataComponent from './components/MetaDataComponent.js';
 
 // Just to ensure that AppStore is initialized and listening to
 // AppDispatcher prior to having router initialized.
 require('./stores/AppStore.js');
 
 /**
- * AppStarter helping to configure and start the app
- *
- * @class  AppStarter
+ * @class
+ * @memberof module:Reactator
+ * @classdesc AppStarter helping to configure and start the app
  */
 class AppStarter {
 
     /**
-     * @method constructor
+     * @constructor
      */
     constructor() {
         this.routes = undefined;
-        this.hash = false;
+        this.history = Router.browserHistory;
     }
 
     /**
      * Sets the hash setting, true to use hash, false to use html5 history
      *
-     * @param {Boolean} useHash to use or not to use
+     * @param {Boolean} history to use or not to use
      * @return {AppStarter} the app starter
-     * @method withHash
      */
-    withHash(useHash) {
-        this.hash = useHash;
+    withHistory(history) {
+        this.history = history;
         return this;
     }
 
     /**
      * Sets the routes to start the app with
      *
-     * @param {Router} routes routes to start the app with
+     * @param {Route} routes routes to start the app with
      * @return {AppStarter} the app starter
-     * @method withRoutes
      */
     withRoutes(routes) {
         this.routes = routes;
@@ -50,32 +49,39 @@ class AppStarter {
     }
 
     /**
+     * Callback for Router updates
+     * @return {undefined}
+     */
+    onUpdate() {
+        RootRouteComponent.emitter.emit(AppConstants.ROUTE_UPDATE);
+    }
+
+    /**
+     * Renders the app
+     * @return {undefined}
+     */
+    render() {
+        this.routes.initialize();
+
+        ReactDOM.render(
+            <Router history={browserHistory} onUpdate={this.onUpdate}>
+                <Route component={RootRouteComponent}>
+                    {this.routes.getRoutes()}
+                </Route>
+            </Router>,
+            document.getElementById('main')
+        );
+    }
+
+    /**
      * Starts the app
-     *
-     * @method start
+     * @return {undefined}
      */
     start() {
-        // The MetaData Component providing common functionality / support.
-        React.render(
-            <MetaDataComponent />,
+        ReactDOM.render(
+            <MetaDataComponent onComponentDidMount={this.render.bind(this)}/>,
             document.getElementById('_md')
         );
-
-        const handler = function(Handler, state) {
-            AppDispatcher.handleRouteAction(state);
-
-            React.render(
-                    <Handler />,
-                    document.getElementById('main')
-                );
-        };
-
-        this.routes.initialize();
-        if (this.hash === true) {
-            Router.run(this.routes.getRoutes(), handler);
-        } else {
-            Router.run(this.routes.getRoutes(), Router.HistoryLocation, handler);
-        }
     }
 }
 
