@@ -1,12 +1,9 @@
-/* global require, module */
-
-const
-    _ = require('lodash'),
-    Q = require('../lib/q.js'),
-    $ = require('../lib/jquery.js'),
-    Client = require('./Client.js'),
-    ClientError = require('./ClientError.js'),
-    ClientResponse = require('./ClientResponse.js');
+import _ from 'lodash';
+import Q from '../lib/q.js';
+import $ from '../lib/jquery.js';
+import Client from './Client.js';
+import ClientError from './ClientError.js';
+import ClientResponse from './ClientResponse.js';
 
 //
 // Helper to return a Q.Promise that returns ClientResponse or ClientError for an
@@ -15,10 +12,12 @@ const
 const promise = ($ajax) => {
     const getHeaders = () => {
         const headers = {};
+
         _.map(
             $ajax.getAllResponseHeaders().trim().split('\n'),
             (header) => {
                 var index = header.indexOf(':');
+
                 headers[header.substr(0, index).trim()] = header.substr(index + 1).trim();
             }
         );
@@ -42,16 +41,15 @@ const ajax = (settings, request) => {
 };
 
 /**
- * Class representing the implementation of the Client interface utilizing the jQuery's ajax.
- *
- * @class AjaxClient
- * @constructor
+ * @class
+ * @classdesc Class representing the implementation of the Client interface utilizing the jQuery's ajax.
+ * @memberof module:Reactator
  */
 class AjaxClient extends Client {
 
     /**
      * @param {String} uri uri to the resource
-     * @method initialize
+     * @return {undefined}
      */
     constructor(uri) {
         super();
@@ -60,56 +58,65 @@ class AjaxClient extends Client {
     }
 
     /**
-     * {{#crossLink "Client/create:method"}}{{/crossLink}}
-     * @method create
+     * @param {Object} item item to create
+     * @param {Object} settings settings for the ajax request
+     * @return {Q.Promise} deferred result of the request
+     * @inheritdoc
      */
     create(item, settings) {
         return ajax(settings, {
-                method : 'POST',
-                url: this.uri,
-                contentType: 'application/json',
-                data: JSON.stringify(item),
-                dataType : 'json'
-            });
+            method: 'POST',
+            url: this.uri,
+            contentType: 'application/json',
+            data: JSON.stringify(item),
+            dataType: 'json'
+        });
     }
 
     /**
-     * {{#crossLink "Client/read:method"}}{{/crossLink}}
-     * @method read
+     * @param {String} id id of the item to read
+     * @param {Object} settings settings for the ajax request
+     * @return {Q.Promise} deferred result of the request
+     * @inheritdoc
      */
     read(id, settings) {
         return ajax(settings, {
-                method : 'GET',
-                url: `${this.uri}/${id}`,
-                dataType : 'json'
-            });
+            method: 'GET',
+            url: `${this.uri}/${id}`,
+            dataType: 'json'
+        });
     }
 
     /**
-     * {{#crossLink "Client/update:method"}}{{/crossLink}}
-     * @method update
+     * @param {String} id id of the item to update
+     * @param {Object} item item to update to
+     * @param {Object} settings settings for the ajax request
+     * @return {Q.Promise} deferred result of the request
+     * @inheritdoc
      */
     update(id, item, settings) {
         return ajax(settings, {
-                method : 'PUT',
-                url: `${this.uri}/${id}`,
-                contentType: 'application/json',
-                data: JSON.stringify(item),
-                dataType : 'json'
-            });
+            method: 'PUT',
+            url: `${this.uri}/${id}`,
+            contentType: 'application/json',
+            data: JSON.stringify(item),
+            dataType: 'json'
+        });
     }
 
     /*jshint -W024 */
     /**
-     * {{#crossLink "Client/delete:method"}}{{/crossLink}}
-     * @method delete
+     * @param {String} id id of the item to delete
+     * @param {Object} settings settings for the ajax request
+     * @return {Q.Promise} deferred result of the request
+     * @inheritdoc
      */
     delete(id, settings) {
         return ajax(settings, {
-                method : 'DELETE',
-                url: `${this.uri}/${id}`,
-                dataType : 'text'
-            });
+            method: 'DELETE',
+            url: `${this.uri}/${id}`,
+            dataType: 'text'
+        });
     }
 }
 
