@@ -1,16 +1,12 @@
-/* global require, module */
-
-const
-    _ = require('lodash');
+import _ from 'lodash';
 
 /**
- * Error thrown by {{#crossLink "Client"}}{{/crossLink}} on failed operations.
- *
  * @param {Number} status http status code of the error
  * @param {String} message message of the error
  * @param {String} request request that failed
- * @class ClientError
- * @constructor
+ * @class
+ * @classdesc Error thrown by {@link module:Reactator.Client} on failed operations.
+ * @memberof module:Reactator
  */
 function ClientError(status, message, request) {
     this.name = 'ClientError';
@@ -21,37 +17,41 @@ function ClientError(status, message, request) {
 
 ClientError.prototype = Object.create(Error.prototype);
 _.extend(ClientError.prototype, {
-    constructor : ClientError,
+    constructor: ClientError,
 
     /**
      * @return {Boolean} is unauthorized error
-     * @method isUnauthorized
+     * @inner
+     * @memberOf ClientError
      */
-    isUnauthorized : function() {
+    isUnauthorized: function() {
         return this.status === 401;
     },
 
     /**
      * @return {Boolean} is forbidden error
-     * @method isForbidden
+     * @inner
+     * @memberOf ClientError
      */
-    isForbidden : function() {
+    isForbidden: function() {
         return this.status === 403;
     },
 
     /**
      * @return {Boolean} is not found
-     * @method isNotFound
+     * @inner
+     * @memberOf ClientError
      */
-    isNotFound : function() {
+    isNotFound: function() {
         return this.status === 404;
     },
 
     /**
      * @return {Boolean} is service not available
-     * @method isServiceUnavailable
+     * @inner
+     * @memberOf ClientError
      */
-    isServiceUnavailable : function() {
+    isServiceUnavailable: function() {
         return this.status === 503;
     }
 });
