@@ -1,12 +1,9 @@
-/* global window, global, require, module */
-
 /**
- * Wrapper for getting jQuery and all its plugins
- *
  * @class jQuery
+ * @classdesc Wrapper for getting jQuery and all its plugins
+ * @memberof module:Reactator
  */
-const
-    jQuery = require('jquery');
+import jQuery from 'jquery';
 
 if(typeof window !== 'undefined')  {
     window.$ = global.$ = global.jQuery = jQuery;
