@@ -1,12 +1,9 @@
-/* global require, module */
-
 /**
- * Wrapper for Q
- *
  * @class Q
+ * @classdesc Wrapper for Q
+ * @memberof module:Reactator
  */
-const
-    Q = require('q');
+import Q from 'q';
 
 Q.longStackSupport = true;
 
