@@ -1,14 +1,12 @@
-/* global require */
-
 require('./reactator/build/GulpTasks.js')
     .initialize(
-        [
-            'reactator/',
-            'demo/',
-        ],
-        require('./reactator.json'),
-        [
-            './node_modules',
-            '.'
-        ]
-    );
+    [
+        'reactator/',
+        'demo/',
+    ],
+    require('./reactator.json'),
+    [
+        './node_modules',
+        '.'
+    ]
+);
