@@ -1,8 +1,4 @@
-/* global jest, describe, beforeEach, require, it, expect, runs, waitsFor, runs */
-
-/* jshint unused:false */
-/* jshint -W097 */
-"use strict";
+/* global jest, describe, beforeEach, it, expect */
 
 jest.autoMockOff();
 jest.mock('../../dispatcher/AppDispatcher.js');
@@ -20,6 +16,7 @@ describe('AppAction', function() {
         AppAction.resize(1,2, AppConstants.LG);
 
         var AppDispatcher = require('../../dispatcher/AppDispatcher.js');
+
         expect(AppDispatcher.handleWindowResize.mock.calls.length).toBe(1);
         expect(AppDispatcher.handleWindowResize.mock.calls[0][0].width).toBe(1);
         expect(AppDispatcher.handleWindowResize.mock.calls[0][0].height).toBe(2);
