@@ -1,17 +1,14 @@
-/* global require, module */
-
 /**
- * Class representing the ClientResponse returned by the {{#crossLink "Client"}}{{/crossLink}}
- *
- * @class ClientResponse
- * @constructor
+ * @class
+ * @classdesc Class representing the ClientResponse returned by the {@link module:Reactator.Client}
+ * @memberof module:Reactator
  */
 class ClientResponse {
     /**
      * @param {Object} value value of the response
      * @param {Number} status status of the response
      * @param {Object} metaData any meta data associated with the response
-     * @method initialize
+     * @constructor
      */
     constructor(value, status, metaData) {
         this.value = value;
