@@ -1,11 +1,9 @@
-/* global jest, describe, afterEach, beforeEach, require, it, expect, runs, waitsFor, runs */
-
-/* jshint newcap:false, unused:false, -W024, -W097 */
-"use strict";
+/* global jest, describe, beforeEach, afterEach, it, expect */
 
 jest.autoMockOff();
 
 var Q = require('../../lib/q.js');
+
 function verify(promise) {
     Q(promise)
         .then(function() {
@@ -13,7 +11,7 @@ function verify(promise) {
         })
         .catch(function(error) {
             expect(error.status).toBe(400);
-            expect(error.message).toBe("Method not implemented!");
+            expect(error.message).toBe('Method not implemented!');
         })
         .done();
 }
