@@ -1,16 +1,11 @@
-/* global require, module */
+import _ from 'lodash';
+import React from 'react';
+import AppStore from '../stores/AppStore.js';
+import AppConstants from '../constants/AppConstants.js';
 
-const
-    _ = require('lodash'),
-    React = require('react'),
-    AppStore = require('../stores/AppStore.js'),
-    AppConstants = require('../constants/AppConstants.js');
-
-/**
- * @method getPropVisibility
- */
 const getPropVisibility = (props, visibility) => {
     var isVisible = false;
+
     return _.find(
         AppConstants.VISIBILITY_ORDER,
         (level) => {
@@ -19,38 +14,41 @@ const getPropVisibility = (props, visibility) => {
         });
 };
 
-/**
- * React component for adapting to the window size and visibility according
- * to {{#crossLink "AppStore"}}{{/crossLink}}. The ResponsiveComponent will render
- * the largest child that would be visible or empty div is non is applicable, e.g.:
- *
+/*
  *     <ResponsiveComponent
  *             MD={<div>You shall render LG or MD!</div>}
  *             XS={<div>You shall render SM or XS!</div>} />
+ */
+
+/**
+ * @class
+ * @classdesc React component for adapting to the window size and visibility according
+ *            to {@link module:Reactator.AppStore}. The ResponsiveComponent will render
+ *            the largest child that would be visible or empty div is non is applicable.
  *
- * @class ResponsiveComponent
+ * @memberof module:Reactator
  */
 const ResponsiveComponent = React.createClass({
 
-    getInitialState : function() {
+    getInitialState: function() {
         return {
-            visibility : AppStore.getState().get('size').get('visibility')
+            visibility: AppStore.getState().get('size').get('visibility')
         };
     },
 
-    componentWillMount : function() {
+    componentWillMount: function() {
         AppStore.addChangeListener(this.onAppStoreChange, 'size');
     },
 
-    componentWillUnmount : function() {
+    componentWillUnmount: function() {
         AppStore.removeChangeListener(this.onAppStoreChange, 'size');
     },
 
-    onAppStoreChange : function() {
+    onAppStoreChange: function() {
         this.setState(this.getInitialState());
     },
 
-    shouldComponentUpdate : function(nextProps, nextState) {
+    shouldComponentUpdate: function(nextProps, nextState) {
         return getPropVisibility(this.props, this.state.visibility) !==
                 getPropVisibility(nextProps, nextState.visibility);
     },
