@@ -1,20 +1,19 @@
-/* global require, module */
+import keyMirror from 'keymirror';
 
 /**
- * All the main App constants that the AppDispatcher or others may use.
- *
- * @class AppConstants
+ * @class
+ * @classdesc All the main App constants that the AppDispatcher or others may use.
+ * @memberof module:Reactator
  */
-const keyMirror = require('keymirror');
-
 const AppConstants = keyMirror({
-  ROUTE_ACTION : null,
-  WINDOW_RESIZE_ACTION : null,
+    ROUTE_UPDATE: null,
+    ROUTE_ACTION: null,
+    WINDOW_RESIZE_ACTION: null,
 
-  XS : null,
-  SM : null,
-  MD : null,
-  LG : null
+    XS: null,
+    SM: null,
+    MD: null,
+    LG: null
 });
 
 AppConstants.VISIBILITY_ORDER = [
