@@ -1,38 +1,36 @@
-/* global require, module */
+import AppConstants from '../constants/AppConstants.js';
+import DispatcherFactory from '../lib/DispatcherFactory.js';
 
 /**
- * Main dispatcher for the App
- *
- * @class AppDispatcher
+ * @class
+ * @classdesc Main dispatcher for the App
+ * @memberof module:Reactator
  */
-const
-    AppConstants = require('../constants/AppConstants.js'),
-    DispatcherFactory = require('../lib/DispatcherFactory.js'),
-    AppDispatcher = DispatcherFactory.create('AppDispatcher');
+const AppDispatcher = DispatcherFactory.create('AppDispatcher');
 
 /**
  * Dispatches a route action of ROUTE_ACTION
  *
- * @method handleRouteAction
  * @param {Object} state route state to be dispatched
+ * @return {undefined}
  */
 AppDispatcher.handleRouteAction = function(state) {
     this.dispatch({
-        source : AppConstants.ROUTE_ACTION,
-        action : state
+        source: AppConstants.ROUTE_ACTION,
+        action: state
     });
 };
 
 /**
  * Dispatches a window resize action of WINDOW_RESIZE_ACTION
  *
- * @method handleWindowResize
  * @param {Object} state window state to be dispatched
+ * @return {undefined}
  */
 AppDispatcher.handleWindowResize = function(state) {
     this.dispatch({
-        source : AppConstants.WINDOW_RESIZE_ACTION,
-        action : state
+        source: AppConstants.WINDOW_RESIZE_ACTION,
+        action: state
     });
 };
 
