@@ -1,13 +1,11 @@
-/* global require, module */
+import {Dispatcher} from 'flux';
 
-const
-    Dispatcher = require('flux').Dispatcher,
-    dispatchers = {};
+const dispatchers = {};
 
 /**
- * Dispatcher factory, ensures creation of single dispatcher for a given name
- *
- * @class DispatcherFactory
+ * @class
+ * @classdesc Dispatcher factory, ensures creation of single dispatcher for a given name
+ * @memberof module:Reactator
  */
 const DispatcherFactory = {
 
@@ -16,14 +14,14 @@ const DispatcherFactory = {
      *
      * @param  {String} name name of the dispatcher (unique)
      * @return {Dispatcher} the newly created dispatcher, or the existing one.
-     * @method create
      */
-    create : function(name) {
+    create: function(name) {
         if (dispatchers[name]) {
             return dispatchers[name];
         }
 
         const dispatcher = new Dispatcher();
+
         dispatcher.name = name;
 
         dispatchers[name] = dispatcher;
